project_vajra/requirements.txt
# Production Dependencies
requests>=2.31.0
python-dateutil>=2.8.2
pandas>=2.0.3
numpy>=1.24.4
python-dotenv>=1.0.0
networkx>=3.2.1  # Added for advanced graph analytics
anthropic>=0.18.0  # Added for Claude threat intelligence
web3>=6.15.0  # Added for Ethereum/Polygon chain analysis
beautifulsoup4>=4.12.3  # Added for AXIOM report parsing
pytsk3>=20210419; sys_platform == "linux"  # Added for FTK image parsing
pyewf>=20220101; sys_platform == "linux"  # Added for EWF image support

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# Async Task Queue
celery[redis]>=5.3.0
redis>=5.0.0

# Forensic Processing
pytsk3>=20210419; sys_platform == "linux"
pyaff4>=0.31; sys_platform == "linux"

# Performance Optimization
uvloop>=0.19.0; sys_platform != "win32"
orjson>=3.9.10

# Security
cryptography>=42.0.5
gnupg>=0.5.1

# Testing (dev only)
pytest>=8.0.0
pytest-cov>=4.1.0
httpx>=0.27.0